back-end/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
alembic>=1.13,<2.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<4.1
python-jose[cryptography]>=3.3,<4.0
pydantic-settings>=2.4,<3.0
email-validator>=2.2,<3.0
